enterprise_service/requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==22.1.0
cached-property==1.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dynaconf==3.1.9
ecdsa==0.18.0
fastapi==0.85.0
greenlet==1.1.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
isodate==0.6.1
lxml==4.9.1
netifaces==0.11.0
packaging==21.3
platformdirs==2.5.2
priority==2.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
wrapt==1.14.1
wsproto==1.2.0
zeep==4.1.0
